requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0          
python-dotenv==1.0.0